+++ src/Sql/Sql/ChangeLog.md
@@ -33,6 +33,8 @@
 * Updated cmdlets `New-AzSqlDatabaseImport` and `New-AzSqlDatabaseExport` to support network isolation functionality
 * Added cmdlet `New-AzSqlDatabaseImportExisting`
 * Update to Databases cmdlets to support backup storage type specification
+* Added Force parameter to `New-AzSqlDatabase`
+* Added warning for BackupStorageRedundancy configuration in select regions in `New-AzSqlDatabase`
 
 ## Version 2.9.1
 * Fixed potential server name case insensitive error in `New-AzSqlServer` and `Set-AzSqlServer`